SimonSays java application
For my Fundamentals of Computer Science 2 class I created a Simon Says Java application. Simon Says is a memory game with four colors where the game generates a random sequence of colors, and the player needs to replay that sequence in the correct order.

To make the game engaging, I implemented a two-phase design where the game generates a sequence of lights and displays it to the player, and then accepts user clicks and checks whether their clicks match the intended sequence. The game gets progressively harder as it lengthens the sequence after each successful playback.

In terms of design, I chose to implement the game as a single subclass of World that has extra fields. I also utilized the Color class's methods to generate easily distinguishable colors for the game's buttons.

The game is designed to tick twice per second, and the player can end the game if they fail to replay the current sequence of colors. The final scene will indicate that the player has lost the game.